# apps/backend/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: NFTSol API
  description: |
    # NFTSol - Revolutionary NFT Marketplace on Solana

    NFTSol is a next-generation NFT marketplace featuring:
    - **Archive Integration**: Access 20M+ public domain items from Internet Archive
    - **Grok AI Verification**: AI-powered authenticity verification for every NFT
    - **Eternal Echoes**: Collaborative NFT layers with contributor rewards
    - **Ultra-Cheap Minting**: Lowest fees on Solana with compressed NFTs

    ## Authentication

    Most endpoints support two authentication methods:
    1. **Wallet Signature**: Sign a message with your Solana wallet
    2. **JWT Token**: Use Bearer token from admin authentication

    All requests should include the `X-CSRF-Token` header for CSRF protection.

    ## Rate Limiting

    - Standard endpoints: 100 requests per 15 minutes per IP
    - Health checks: Unlimited
    - Admin endpoints: Requires JWT token

    ## Response Format

    All responses follow a consistent format:

    ```json
    {
      "success": true,
      "data": { "...": "..." },
      "message": "Optional message",
      "requestId": "Unique request ID"
    }
    ```

    Error responses include an error code for better error handling.
  version: 1.0.0
  contact:
    name: NFTSol Support
    url: https://github.com/TheoryofShadows/nftsol
    email: support@nftsol.app
  license:
    name: MIT
    url: https://github.com/TheoryofShadows/nftsol/blob/main/LICENSE
servers:
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from admin authentication
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF protection token
security:
  - csrfToken: []
paths:
...
  /api/v1/simple-mint:
    post:
      tags:
        - NFTs
      summary: Mint NFT
      description: Create and mint a new NFT on Solana
      operationId: mintNft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: NFT name
                description:
                  type: string
                  description: NFT description
                imageUrl:
                  type: string
                  format: uri
                  description: Image URL for the NFT
                creatorWallet:
                  type: string
                  description: Creator wallet address
              required:
                - name
                - imageUrl
            examples:
              basic:
                summary: Basic NFT mint
                value:
                  name: "My First NFT"
                  description: "A beautiful digital artwork"
                  imageUrl: "https://example.com/image.jpg"
                  creatorWallet: "11111111111111111111111111111111"
      responses:
        '200':
          description: NFT minted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      mintAddress:
                        type: string
                      transactionSignature:
                        type: string
                      name:
                        type: string
        '400':
          description: Validation error
        '503':
          description: Solana network unavailable
...
  /api/echo/mint:
    post:
      tags:
        - Echo (Collaborative NFTs)
      summary: Create Echo NFT
      description: Create a new Eternal Echo (collaborative NFT) with layers
      operationId: createEcho
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                baseMintAddress:
                  type: string
                  description: Base NFT mint address
                creatorWallet:
                  type: string
                  description: Creator wallet address
                metadata:
                  type: object
                  description: Echo metadata
              required:
                - baseMintAddress
                - creatorWallet
      responses:
        '200':
          description: Echo created successfully
        '400':
          description: Validation error
...